# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: MCP Server for coding workflows — file editing, code execution, git, package management, and more
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /files:
    get:
      operationId: listFiles
      tags: [files]
      summary: List files in a directory
      parameters:
        - name: path
          in: query
          required: false
          schema:
            type: string
            default: "."
        - name: recursive
          in: query
          required: false
          schema:
            type: boolean
            default: false
        - name: pattern
          in: query
          required: false
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: File listing
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileEntry"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /files/read:
    get:
      operationId: readFile
      tags: [files]
      summary: Read file contents
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
        - name: encoding
          in: query
          required: false
          schema:
            type: string
            default: utf-8
      responses:
        "200":
          description: File content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileContent"
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /files/write:
    post:
      operationId: writeFile
      tags: [files]
      summary: Write or update a file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileWrite"
      responses:
        "200":
          description: File written
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileEntry"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /files/delete:
    delete:
      operationId: deleteFile
      tags: [files]
      summary: Delete a file or directory
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
        - name: recursive
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: File deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /files/search:
    get:
      operationId: searchFiles
      tags: [files]
      summary: Search file contents recursively
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
        - name: path
          in: query
          required: false
          schema:
            type: string
            default: "."
        - name: filePattern
          in: query
          required: false
          schema:
            type: ["string", "null"]
        - name: caseSensitive
          in: query
          required: false
          schema:
            type: boolean
            default: false
        - name: maxResults
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SearchResult"
  /files/move:
    post:
      operationId: moveFile
      tags: [files]
      summary: Move or rename a file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileMove"
      responses:
        "200":
          description: File moved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileEntry"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ---- Execution ----
...
  /git/status:
    get:
      operationId: gitStatus
      tags: [git]
      summary: Get git repository status
      parameters:
        - name: path
          in: query
          required: false
          schema:
            type: string
            default: "."
      responses:
        "200":
          description: Git status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GitStatus"
  /git/log:
    get:
      operationId: gitLog
      tags: [git]
      summary: Get commit history
      parameters:
        - name: path
          in: query
          required: false
          schema:
            type: string
            default: "."
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: branch
          in: query
          required: false
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Commit log
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GitCommit"
  /git/branches:
    get:
      operationId: listGitBranches
      tags: [git]
      summary: List git branches
      parameters:
        - name: path
          in: query
          required: false
          schema:
            type: string
            default: "."
      responses:
        "200":
          description: Branch list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GitBranch"
  /git/commit:
    post:
      operationId: gitCommit
      tags: [git]
      summary: Stage and commit changes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GitCommitInput"
      responses:
        "200":
          description: Commit result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GitCommit"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /git/diff:
    get:
      operationId: gitDiff
      tags: [git]
      summary: Get diff of changes
      parameters:
        - name: path
          in: query
          required: false
          schema:
            type: string
            default: "."
        - name: staged
          in: query
          required: false
          schema:
            type: boolean
            default: false
        - name: file
          in: query
          required: false
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Diff content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GitDiff"

  # ---- Packages ----
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: ["string", "null"]

    # Files
    FileEntry:
      type: object
      required: [name, path, type, size, modifiedAt]
      properties:
        name:
          type: string
        path:
          type: string
        type:
          type: string
          enum: [file, directory, symlink]
        size:
          type: integer
        modifiedAt:
          type: string
        permissions:
          type: ["string", "null"]
    FileContent:
      type: object
      required: [path, content, size, encoding]
      properties:
        path:
          type: string
        content:
          type: string
        size:
          type: integer
        encoding:
          type: string
        mimeType:
          type: ["string", "null"]
    FileWrite:
      type: object
      required: [path, content]
      properties:
        path:
          type: string
        content:
          type: string
        encoding:
          type: string
          default: utf-8
        createDirs:
          type: boolean
          default: true
    FileMove:
      type: object
      required: [from, to]
      properties:
        from:
          type: string
        to:
          type: string
        overwrite:
          type: boolean
          default: false
    SearchResult:
      type: object
      required: [file, line, column, match, context]
      properties:
        file:
          type: string
        line:
          type: integer
        column:
          type: integer
        match:
          type: string
        context:
          type: string

    # Execution
...
    GitStatus:
      type: object
      required: [branch, clean, staged, unstaged, untracked]
      properties:
        branch:
          type: string
        clean:
          type: boolean
        ahead:
          type: integer
        behind:
          type: integer
        staged:
          type: array
          items:
            type: string
        unstaged:
          type: array
          items:
            type: string
        untracked:
          type: array
          items:
            type: string
    GitCommit:
      type: object
      required: [hash, message, author, email, date]
      properties:
        hash:
          type: string
        message:
          type: string
        author:
          type: string
        email:
          type: string
        date:
          type: string
    GitBranch:
      type: object
      required: [name, current, remote]
      properties:
        name:
          type: string
        current:
          type: boolean
        remote:
          type: ["string", "null"]
        lastCommit:
          type: ["string", "null"]
    GitCommitInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
        files:
          type: array
          items:
            type: string
        all:
          type: boolean
          default: false
    GitDiff:
      type: object
      required: [diff, filesChanged, insertions, deletions]
      properties:
        diff:
          type: string
        filesChanged:
          type: integer
        insertions:
          type: integer
        deletions:
          type: integer

    # Packages
...